domain: add --pending-deprecation warning
Moves the domain module implementation to internal/domain.js so
that core can continue to use it without triggering any dep
warnings.

Add a pending deprecation warning that only emits when the
--pending-deprecation flag is used.
