+++ lib/events.js
@@ -62,7 +62,7 @@ EventEmitter.init = function() {
   this.domain = null;
   if (EventEmitter.usingDomains) {
     // if there is an active domain, then attach to it.
-    domain = domain || require('domain');
+    domain = domain || require('internal/domain');
     if (domain.active && !(this instanceof domain.Domain)) {
       this.domain = domain.active;
     }
