+++ lib/repl.js
@@ -54,7 +54,7 @@ const fs = require('fs');
 const Interface = require('readline').Interface;
 const Console = require('console').Console;
 const Module = require('module');
-const domain = require('domain');
+const domain = require('internal/domain');
 const debug = util.debuglog('repl');
 
 const parentModule = module;
